pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "takeout-rater"
version = "0.1.0"
description = "Aesthetics scoring orchestrator for Google Photos Takeout folders"
readme = "README.md"
license = { text = "GPL-3.0-only" }
requires-python = ">=3.12"
dependencies = [
    # Core image handling
    "Pillow>=10.0",
    # Web UI
    "fastapi>=0.100",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
    "aiofiles>=23.0",
    # Aesthetic scorer (LAION Aesthetic Predictor v2)
    "torch>=2.2",
    "torchvision>=0.17",
    "open-clip-torch>=2.24",
    "huggingface-hub>=0.20",
    # NSFW scorer
    "transformers>=4.35",
    # HEIC image support
    "pillow-heif>=0.16",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "httpx>=0.26",
]
brisque = [
    "scikit-image>=0.21",
    "scipy>=1.10",
]

[project.scripts]
takeout-rater = "takeout_rater.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
